+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - added support for multiple code change requests in Updated Files modal
 - added file sidebar in the Updated Files modal
 - improved handling of 'unknown' finish reason of some models when using Agent
+- added file viewer modal for Project Files
+- using no-verify during the rebase to avoid failure
+- added merge conflict indicators for updated files
+- using correct API endpoint for /resolve-conflicts from browser client
+- added API endpoint for getting Aider connector installation status from browser client
 
 ## [0.61.1]